fix: renaming and move table row files
