+++ Source/wunthshin/Data/CharacterTableRow/CharacterTableRow.cpp
