+++ Source/wunthshin/Data/Elements/Crystalize/O_WSCrystalize.h
@@ -3,14 +3,14 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "wunthshin/Data/Elements/O_ElementReactor.h"
+#include "wunthshin/Data/Elements/O_WSElementReactor.h"
 #include "O_WSCrystalize.generated.h"
 
 /**
  * 
  */
 UCLASS()
-class WUNTHSHIN_API UO_WSCrystalize : public UO_ElementReactor
+class WUNTHSHIN_API UO_WSCrystalize : public UO_WSElementReactor
 {
 	GENERATED_BODY()
 };
